Set up basic skeleton for Dexie.js implementation
