+++ packages/browser-lite/package.json
@@ -18,8 +18,10 @@
         "@0x/order-utils": "^10.2.0",
         "@0x/utils": "^5.4.0",
         "ajv": "^6.12.2",
+        "@types/dexie": "^1.3.1",
         "base64-arraybuffer": "^0.2.0",
         "browserfs": "^1.4.3",
+        "dexie": "^3.0.1",
         "ethereum-types": "^3.0.0"
     },
     "devDependencies": {
