+++ packages/test-wasm/package.json
@@ -12,9 +12,11 @@
     },
     "dependencies": {
         "@0x/mesh-browser-lite": "1.0.0",
+        "@types/dexie": "^1.3.1",
+        "dexie": "^3.0.1",
         "shx": "^0.3.2",
-        "typescript": "^3.9.3",
         "ts-loader": "^6.2.1",
+        "typescript": "^3.9.3",
         "webpack": "^4.43.0",
         "webpack-cli": "^3.3.10"
     }
